Renamed and moved StringHelper
