+++ src/main/kotlin/ConfigManager.kt
@@ -1,7 +1,7 @@
 package org.kamiblue.botkt
 
 import com.google.gson.Gson
-import org.kamiblue.botkt.helpers.StringHelper.isUrl
+import org.kamiblue.botkt.utils.StringUtils.isUrl
 import java.io.File
 import java.net.URL
 import java.nio.file.Files
