+++ src/main/kotlin/commands/BanCommand.kt
@@ -16,7 +16,7 @@ import org.kamiblue.botkt.ConfigManager.readConfigSafe
 import org.kamiblue.botkt.Permissions.hasPermission
 import org.kamiblue.botkt.utils.MessageSendUtils.error
 import org.kamiblue.botkt.utils.MessageSendUtils.normal
-import org.kamiblue.botkt.helpers.StringHelper.flat
+import org.kamiblue.botkt.utils.StringUtils.flat
 import org.kamiblue.botkt.utils.Colors
 import org.kamiblue.botkt.utils.ReactionUtils.FakeUser
